The State stores a States object as id, not a string
Instead of just storing a string as `id` attribute, store an actual
States enum object. This means that to get to a textual id, the value from
the enum must be retrieved, or the name used. But it allows for access
to the actual enum, which we use for statistics.