Introduce handler objects to parser that handle output and control
In order to customize how an application reacts to certain input, or
rather how it generates output from it, add handlers to the VT500Parser
that are sent the respective codes and controls. We could have done this
via deriving from the base parser, but this way has two benefits.

One is that the derived class' methods do not remember to call the base
class methods, which we want to keep statistics, etc.
The second is that generating output is something separate from parsing the
input. And since I want to add parsing of lines to find certain sections
and commands in the console log, that is rather something to be done
in a derived class. Having handlers for output makes that less messy.