Add a parser that finds the prompt in a terminal session log
Add `terminalparser.py` which has a class `TermLogParser` which is
derived from the `VT500Parser`. This parser reads the file line by line
and searches for when the prompt comes up and user input is entered.

It emits an event to a configurable event handler when the prompt is
entered and exited. A handling application can the e.g. process user
input entered after the prompt differently than program output.

Apparently it is necessary for every module to have its own logger
instance, defined on module scope.