Redefine EventListener prompt_start to prompt_active.
In order to be able to react to an upcoming prompt, e.g. in the
conversion to HTML, we need an event for that. The `prompt_start`
was already used but I couldn't think of a better name.
So the existing `prompt_start` is renamed to `prompt_active`. That
describes it better, since the state is that the prompt has been written
and now the command input is handled.
The `prompt_start` method is now a new event that signals that the
prompt, i.e. the prompt string, start now.