Add option to terminalreplay to suppress Vim sessions
The mamber variable `print_vim` can be set to `False`
in order to suppress output when a Vim session is active.

There still seems to be a not quite understood case, where
the last line of a vim session is still put out. This can
result in short screen blinking.