The emit method is alone responsible to determine vim_end event causes
The emit method already has a block under the STATE_NORMAL handling
where an ongoing vim session is ended. But the code for checking for
a vim session when a OSC comes in was in the `osc_end` method.

To keep the logic together, the code is moved to the `emit` method,
where it is now handled just like for the STATE_NORMAL signal.