terminalparser: emit a state change event before changing the state
When changing to a new TLP state, first emit an event, then change to
the new state. This lets the `emit` function send the correct event
depending on the current state, before the new state gets active.
Otherwise the information about the state previous to the state change
would already be lost.