terminalparser: Rewrite the detection of Vim editor sessions
The Vim editor is not so easy to detect. The first approach does not
catch all sessions, the detection of terminal behaviour is not always done.
Maybe some older Vim versions don't.
But in general Vim behaves differently in our terminal session logs.
So the detection is rewritten, searching for the hint that Git gives us,
and for the xterm window manipulation controls sequences "[22;0;0t" and
"[22;2t". The "[22;0;0t" alone would not suffice, since we see 'less'
using that, too.