toHtml: Handle 39m, 49m, and 22m CSI controls sequences
So far we ignored them because they go back to the default color, or
not bold. Instead of opening another span it would be better to close
an open one. But we only closed all spans on a reset control sequence.

Now we keep a stack of spans and if, e.g. a foreground color should be
set back to default, we close all spans that came after the one we want
closed, close the foreground color one, and then open the ones that
we just closed want which should still be in effect.