Find prompt without OSC
I removed the OSC to set the window title from the prompt. Bad idea.
I didn't remember that I use it in the parser to find the prompt. So
now a method is required to find a prompt on a line. Not just without
OSC but also in logs originating from a Linux terminal instead of the
Windows Git terminal, where the prompt sits on its extra line.

In addition, git commands may enable application mode in the terminal,
which leads to the commands for disabling it to be on the line with
the new prompt, too.