Detect prompts in line after application mode switches
Detecting a prompt without a OSC is not that easy. A line may not start
with the prompt directly, and the prompt may come after other output
on a line without a CRLF first. Try to detect this for cases where
commands like "git" or "man" switches to application mode for output,
and ends it before returning to the prompt.