Fix CarriageReturn handling in command line
A CR in the command line (i.e. when we type) will return to the start
of the line, i.e. including the prompt. Stepping through history can
result in such control, followed by moving the cursor to the right again.

To account for this, the command line builder must know how many
characters there are in the prompt before the command line (the typing)
and walk over them when a CR is handled.

To hanlde this, a command line LineBuilder gets the (printable) length
of the LineBuilder preceding it. A CR will make the current position
go negative if there are characters in the line before the command.
The escape and CSI function handling and printing is adapted accordingly.