End vim session recording at frame that includes the session end
Since our asciinema recording parser assembles a complete line it
will put too much asciinema frames into the vim session recording.
This is because a newline may come only much later.
Ideally we end the session recording before the vim buffer is closed,
so that the contents are still visible in the last frame.
For this, search each frame for the end of the vim sessions. We simply
reuse the regular expressions defined in TerminalParser. When we find
the vim session end, we stop adding frames to the session recording
right there.