Pass vim start event info (height, file name) with event
We already detect (for the log) in which height a vim session starts
and for which file, if that information is detected in the terminal
session.
This is useful for asciinema recordings of vim sessions, but was so
far not available anywhere else in the program.
So extend the `vim_start` event with properties which are passed with
the event to the event handlers. This way it can be used in the
`asciinema2html` parser. The two properties passed along are currently
`height` and `file`.