Detect vim session height and file name also in one-line session
We detect when a second vim session is opened on the same line, but
so far did not make the effort to detect size and file name properties
for this one. Maybe because the chnaces are slim that something changed
in between, at least regarding the height.
But why not also try to get the information if it is available. We are
not trying to win a speed race here.