Lower logging levels for info and debug in vtparser
Let's lower the level from info to debug and from debug to trace in
vtparser. The trace level is introduced as a self defined level.
This way we can make use of them when running the vtparser module
directly, but skip them when run from a different module.

For this, the logger in vtparser get explicitly named as `vtparser`
so that its level can be adjusted from the outside.