html: Switch to flex layout for command row
The command row used floats to get the command number to the left of
the command. This comes with all the problems of this legacy layout.
Change this to use flexbox for the command row, which places the
command to the right of the number, without wrapping under the number
when the command gets too big.

This needs the introduction of a command wrapper to make sure that
commands with multiple box elements flow in cross direction in the
future.