Detect vim session end if frame was already added to frame buffer
When the start of a vim session is only detected at the end of the vim
session, because there was no newline before it, then the frame that
includes the end may end up in the buffered frames. Which makes
the session recoridng longer than it should be.
So check if the last frame in the buffered list of frames includes the
end of the vim session. If so, drop it and stop the capture.
This only looks at the last frame. That might not be enough, it only
fixes the one problem I currently see.