Add option "ahopto" to file list which turns on automatic hop detection for review
The file list file can have a "ahopto" member for a file. If this is
set to the id of the hop target file, then a hop detection is run on
the two files. The result is that in review mode you get detected hops
to the respective time stamp in the other file listed in the output.
This helps when putting together the hopto list. It is not a fully
automatic hop listing. This is also because maybe the detected hop
target is not the ideal command for the flow.